internal/engine/detect: extract the provider-generic detection engine
Move the detection rule-DSL engine out of package github into a shared
package internal/engine/detect: the getPath field resolver, the
where-clause evaluator and all_of/any_of/none_of combinators, chain_of
evaluation, the YAML rule loader, evidence templating, and the scan
driver. The logic was already provider-generic; it is now parameterized
by a Provider seam — the subject-kind to normalize-dir map, the rule
subtree, the provider label, and the finding-construction hooks
(Display/Code/Repo/File/SubjectKey).

GitHub becomes a thin adapter: its subjectDisplay/buildCode and
subject-dir map move verbatim into a github Provider, and
github.Scan/ScanOptions are kept so the CLI is unchanged. GitHub
findings are byte-identical — getpath/evaluate/evidence are unchanged
and the full github suite passes.

SubjectKey gives platforms whose subjects lack a natural _id (e.g. ADO's
derived-edge records) a stable, unique finding identity; nil falls back
to SubjectHash so github is unaffected.
