+++ internal/engine/detect/evaluate.go
@@ -1,4 +1,4 @@
-package github
+package detect
 
 import (
 	"fmt"
