+++ internal/engine/detect/getpath.go
@@ -1,4 +1,4 @@
-package github
+package detect
 
 import "strings"
