+++ internal/engine/detect/getpath_test.go
@@ -1,4 +1,4 @@
-package github
+package detect
 
 import (
 	"reflect"
