+++ internal/github/orgonly_test.go
@@ -3,10 +3,12 @@ package github
 import (
 	"slices"
 	"testing"
+
+	"github.com/praetorian-inc/trajan/internal/engine/detect"
 )
 
 func TestOrgOnlyFilterSelectsExactlyOrgSubjects(t *testing.T) {
-	rules, err := LoadRules()
+	rules, err := detect.LoadRules("github")
 	if err != nil {
 		t.Fatalf("LoadRules: %v", err)
 	}
@@ -14,7 +16,7 @@ func TestOrgOnlyFilterSelectsExactlyOrgSubjects(t *testing.T) {
 	full := len(rules)
 	// exercise the production filter (--org-detections-only path) directly; clone
 	// first because orgOnlyRules mutates its input via slices.DeleteFunc.
-	org := orgOnlyRules(slices.Clone(rules))
+	org := detect.OrgOnlyRules(slices.Clone(rules))
 	if len(org) == 0 {
 		t.Fatal("expected at least one subject==org rule")
 	}
@@ -40,11 +42,11 @@ func TestOrgOnlyFilterSelectsExactlyOrgSubjects(t *testing.T) {
 }
 
 func TestMembersCanCreatePrivateReposRuleFires(t *testing.T) {
-	rules, err := LoadRules()
+	rules, err := detect.LoadRules("github")
 	if err != nil {
 		t.Fatalf("LoadRules: %v", err)
 	}
-	var rule *Rule
+	var rule *detect.Rule
 	for i := range rules {
 		if rules[i].ID == "cat-13/members-can-create-private-repositories" {
 			rule = &rules[i]
@@ -59,12 +61,12 @@ func TestMembersCanCreatePrivateReposRuleFires(t *testing.T) {
 	}
 
 	fires := map[string]any{"_id": "acme", "members_can_create_private_repositories": true}
-	if got := EvaluateRule(rule, []map[string]any{fires}, nil); len(got) != 1 {
+	if got := detect.EvaluateRule(rule, []map[string]any{fires}, nil); len(got) != 1 {
 		t.Errorf("rule should fire when members_can_create_private_repositories is true, got %d", len(got))
 	}
 
 	silent := map[string]any{"_id": "acme", "members_can_create_private_repositories": false}
-	if got := EvaluateRule(rule, []map[string]any{silent}, nil); len(got) != 0 {
+	if got := detect.EvaluateRule(rule, []map[string]any{silent}, nil); len(got) != 0 {
 		t.Errorf("rule should stay silent when the setting is false, got %d", len(got))
 	}
 }
